+++ examples/default-provider/app/api/video/route.js
@@ -1,0 +1,1 @@
+export { GET } from 'next-video/request-handler';