+++ examples/default-provider/app/layout.tsx
@@ -47,7 +47,13 @@ export default async function RootLayout({
                 <Link href="/">Basic example</Link>
               </li>
               <li>
+                <Link href="/background-video">Background Video</Link>
+              </li>
+              <li>
                 <Link href="/slotted-poster">Slotted Poster</Link>
+              </li>
+              <li>
+                <Link href="/string-source">String video source</Link>
               </li>
             </ul>
           </nav>