+++ src/webpack/video-raw-loader.ts
@@ -1,12 +1,18 @@
 import path from 'node:path';
 import { readFile } from 'node:fs/promises';
-import { createAsset, getAssetConfigPath } from './assets.js';
+import { createAsset, getAssetConfigPath } from '../assets.js';
+
 
 // https://webpack.js.org/api/loaders#raw-loader
 export const raw = true;
 
-export default async function loader(this: any, source: Buffer) {
-  const importPath = `${this.resourcePath}${this.resourceQuery ?? ''}`;
+export default async function loader(this: any) {
+  const params = new URLSearchParams(this.resourceQuery);
+  // Don't include the thumbnailTime URL param in the asset path.
+  params.delete('thumbnailTime');
+
+  const query = params.size ? `?${params}` : '';
+  const importPath = `${this.resourcePath}${query}`;
   const assetPath = path.resolve(await getAssetConfigPath(importPath));
 
   this.addDependency(assetPath);
@@ -19,6 +25,5 @@ export default async function loader(this: any, source: Buffer) {
       status: 'sourced'
     }));
   }
-
-  return `export default ${asset}`;
+  return `${asset}`;
 }